+++ src/resources/addressing/address-maps/address-maps.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import * as AddressMapsAPI from './address-maps';
 import * as IPsAPI from '../../ips';
@@ -35,10 +36,18 @@ export class AddressMaps extends APIResource {
    * ```
    */
   create(
-    params: AddressMapCreateParams,
+    params?: AddressMapCreateParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<AddressMapCreateResponse>;
+  create(options?: Core.RequestOptions): Core.APIPromise<AddressMapCreateResponse>;
+  create(
+    params: AddressMapCreateParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<AddressMapCreateResponse> {
-    const { account_id, ...body } = params;
+    if (isRequestOptions(params)) {
+      return this.create({}, params);
+    }
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/addressing/address_maps`, {
         body,
@@ -61,10 +70,18 @@ export class AddressMaps extends APIResource {
    * ```
    */
   list(
-    params: AddressMapListParams,
+    params?: AddressMapListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<AddressMapsSinglePage, AddressMap>;
+  list(options?: Core.RequestOptions): Core.PagePromise<AddressMapsSinglePage, AddressMap>;
+  list(
+    params: AddressMapListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<AddressMapsSinglePage, AddressMap> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return this._client.getAPIList(
       `/accounts/${account_id}/addressing/address_maps`,
       AddressMapsSinglePage,
@@ -87,10 +104,19 @@ export class AddressMaps extends APIResource {
    */
   delete(
     addressMapId: string,
-    params: AddressMapDeleteParams,
+    params?: AddressMapDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<AddressMapDeleteResponse>;
+  delete(addressMapId: string, options?: Core.RequestOptions): Core.APIPromise<AddressMapDeleteResponse>;
+  delete(
+    addressMapId: string,
+    params: AddressMapDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<AddressMapDeleteResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(addressMapId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return this._client.delete(`/accounts/${account_id}/addressing/address_maps/${addressMapId}`, options);
   }
 
@@ -110,7 +136,7 @@ export class AddressMaps extends APIResource {
     params: AddressMapEditParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<AddressMap> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.patch(`/accounts/${account_id}/addressing/address_maps/${addressMapId}`, {
         body,
@@ -132,10 +158,19 @@ export class AddressMaps extends APIResource {
    */
   get(
     addressMapId: string,
-    params: AddressMapGetParams,
+    params?: AddressMapGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<AddressMapGetResponse>;
+  get(addressMapId: string, options?: Core.RequestOptions): Core.APIPromise<AddressMapGetResponse>;
+  get(
+    addressMapId: string,
+    params: AddressMapGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<AddressMapGetResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(addressMapId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(
         `/accounts/${account_id}/addressing/address_maps/${addressMapId}`,
@@ -431,7 +466,7 @@ export interface AddressMapCreateParams {
   /**
    * Path param: Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: An optional description field which may be used to describe the
@@ -475,21 +510,21 @@ export interface AddressMapListParams {
   /**
    * Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface AddressMapDeleteParams {
   /**
    * Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface AddressMapEditParams {
   /**
    * Path param: Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: If you have legacy TLS clients which do not send the TLS server name
@@ -517,7 +552,7 @@ export interface AddressMapGetParams {
   /**
    * Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 AddressMaps.AddressMapsSinglePage = AddressMapsSinglePage;
