+++ src/resources/api-gateway/labels/user/user.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../../resource';
+import { isRequestOptions } from '../../../../core';
 import * as Core from '../../../../core';
 import * as ResourcesAPI from './resources/resources';
 import { Resources } from './resources/resources';
@@ -25,7 +26,7 @@ export class User extends APIResource {
     params: UserUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<UserUpdateResponse> {
-    const { zone_id, ...body } = params;
+    const { zone_id = this._client.zoneId, ...body } = params;
     return (
       this._client.put(`/zones/${zone_id}/api_gateway/labels/user/${name}`, {
         body,
@@ -47,10 +48,19 @@ export class User extends APIResource {
    */
   delete(
     name: string,
-    params: UserDeleteParams,
+    params?: UserDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<UserDeleteResponse>;
+  delete(name: string, options?: Core.RequestOptions): Core.APIPromise<UserDeleteResponse>;
+  delete(
+    name: string,
+    params: UserDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<UserDeleteResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(name, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return (
       this._client.delete(`/zones/${zone_id}/api_gateway/labels/user/${name}`, options) as Core.APIPromise<{
         result: UserDeleteResponse;
@@ -78,7 +88,7 @@ export class User extends APIResource {
     params: UserBulkCreateParams,
     options?: Core.RequestOptions,
   ): Core.PagePromise<UserBulkCreateResponsesSinglePage, UserBulkCreateResponse> {
-    const { zone_id, body } = params;
+    const { zone_id = this._client.zoneId, body } = params;
     return this._client.getAPIList(
       `/zones/${zone_id}/api_gateway/labels/user`,
       UserBulkCreateResponsesSinglePage,
@@ -100,10 +110,20 @@ export class User extends APIResource {
    * ```
    */
   bulkDelete(
-    params: UserBulkDeleteParams,
+    params?: UserBulkDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<UserBulkDeleteResponsesSinglePage, UserBulkDeleteResponse>;
+  bulkDelete(
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<UserBulkDeleteResponsesSinglePage, UserBulkDeleteResponse>;
+  bulkDelete(
+    params: UserBulkDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<UserBulkDeleteResponsesSinglePage, UserBulkDeleteResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.bulkDelete({}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return this._client.getAPIList(
       `/zones/${zone_id}/api_gateway/labels/user`,
       UserBulkDeleteResponsesSinglePage,
@@ -127,7 +147,7 @@ export class User extends APIResource {
     params: UserEditParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<UserEditResponse> {
-    const { zone_id, ...body } = params;
+    const { zone_id = this._client.zoneId, ...body } = params;
     return (
       this._client.patch(`/zones/${zone_id}/api_gateway/labels/user/${name}`, {
         body,
@@ -147,8 +167,17 @@ export class User extends APIResource {
    * );
    * ```
    */
-  get(name: string, params: UserGetParams, options?: Core.RequestOptions): Core.APIPromise<UserGetResponse> {
-    const { zone_id, ...query } = params;
+  get(name: string, params?: UserGetParams, options?: Core.RequestOptions): Core.APIPromise<UserGetResponse>;
+  get(name: string, options?: Core.RequestOptions): Core.APIPromise<UserGetResponse>;
+  get(
+    name: string,
+    params: UserGetParams | Core.RequestOptions = {},
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<UserGetResponse> {
+    if (isRequestOptions(params)) {
+      return this.get(name, {}, params);
+    }
+    const { zone_id = this._client.zoneId, ...query } = params;
     return (
       this._client.get(`/zones/${zone_id}/api_gateway/labels/user/${name}`, {
         query,
@@ -333,7 +362,7 @@ export interface UserUpdateParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param: The description of the label
@@ -350,14 +379,14 @@ export interface UserDeleteParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface UserBulkCreateParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param
@@ -388,14 +417,14 @@ export interface UserBulkDeleteParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface UserEditParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param: The description of the label
@@ -412,7 +441,7 @@ export interface UserGetParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Query param: Include `mapped_resources` for each label
